requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
paramiko>=3.4.0
PyYAML>=6.0.1
PySide6>=6.6.0
